Refactoring to organize components
